test(agentos): the cockpit container spec splits into ten concern-named siblings; the shared fakes move to one module (#97)
A pure move along the catch-all's ten describe boundaries: 108 arms byte-identical (a slicing script re-read every output and asserted it), each file with its block's JSDoc, its own setup app name, the same engine environment and only the fakes it drives. cockpitFakes.mjs holds the five shared fakes and the seed path, exported by name. Two blocks never had a JSDoc (whole-fleet control, seat-identity posture); theirs is written from the arm titles. Unit tree 758/758, dev's count.
